agents/resourceplanner_adk/requirements.txt
google-cloud-firestore
fastapi
uvicorn
requests
adk